feat: add pagina da planta - Co-authored-by: martinsjeniffer
