+++ pages/NovaPlanta/step-1.tsx
@@ -7,6 +7,8 @@ import Main from "../../components/Main";
 import { CardSelect, Content, ContentForm, ContentFormSM, DivCards, NovaPlantaHeader, NovaPlantaText, SubmitDiv } from "../../styles/NovaPlanta";
 import Input from "../../components/Input";
 import Button from "../../components/Button";
+import BottomMenu from "../../components/BottomMenu";
+import { useRouter } from "next/router";
 
 type Inputs = {
   nome: string,
@@ -15,7 +17,7 @@ type Inputs = {
   senha: string,
 };
 
-export default function Step3(props) {
+export default function Step1(props) {
   const {
     register,
     handleSubmit,
@@ -24,59 +26,58 @@ export default function Step3(props) {
   } = useForm<Inputs>();
   
   const onSubmit: SubmitHandler<Inputs> = data => console.log(data);
-  
+  const router = useRouter();
+
   console.log(watch("email"), errors)
   
   return (
     <Main backgroundColor="#FFFFFF" padding="2">
         <NovaPlantaHeader>
-          <h2>Selecione sua nova <span>planta</span></h2>
-          <h3>Confira o que temos para você:</h3>
+          <h1>Sua nova <span>planta</span></h1>
         </NovaPlantaHeader>
           <ContentFormSM onSubmit={handleSubmit(onSubmit)}>
-            <Content className="contentsm">
+            <Content>
+              <NovaPlantaText>Você está em:</NovaPlantaText>
+                <Input
+                  {...register("email", { required: true })}
+                  placeholder="Avenida Paulista, 1234"
+                  border="1px solid #9AC979"
+                  backgroundColor="#9AC979"
+                  color="#FFF"
+                  width={100}
+                  margin="0"
+                />
             <div>
+              <NovaPlantaText>Quanto espaço você tem para plantar?</NovaPlantaText>
               <DivCards>
                 <CardSelect>
-                  <Image src="/tipo-planta/alface.png" alt="Alface" width={95} height={95} />
-                  <p>Alface</p>
+                  <Image src="/planta_espaco_pequeno.png" alt="Pequeno" width={95} height={95} />
+                  <p>Pequeno</p>
                 </CardSelect>
                 <CardSelect>
-                  <Image src="/tipo-planta/pepino.png" alt="Pepino" width={95} height={95} />
-                  <p>Pepino</p>
+                  <Image src="/planta_espaco_medio.png" alt="Médio" width={95} height={95} />
+                  <p>Médio</p>
                 </CardSelect>
                 <CardSelect>
-                  <Image src="/tipo-planta/morango.png" alt="Morango" width={95} height={95} />
-                  <p>Morango</p>
-                </CardSelect>
-            </DivCards>
-            <DivCards>
-                <CardSelect>
-                  <Image src="/tipo-planta/rabanete.png" alt="Rabanete" width={95} height={95} />
-                  <p>Rabanete</p>
-                </CardSelect>
-                <CardSelect>
-                  <Image src="/tipo-planta/manjericao.png" alt="Manjericão" width={95} height={95} />
-                  <p>Manjericão</p>
-                </CardSelect>
-                <CardSelect>
-                  <Image src="/tipo-planta/alho.png" alt="Alho" width={95} height={95} />
-                  <p>Alho</p>
+                  <Image src="/planta_espaco_grande.png" alt="Grande" width={95} height={95} />
+                  <p>Grande</p>
                 </CardSelect>
               </DivCards>
             </div>
             
-            <SubmitDiv className="contentsm">
+            <SubmitDiv>
               <Button
                 backgroundColor="#217B51"
                 width="min-content"
                 color="#FFF"
                 padding="15px 100px"
                 fontSize="18px"
+                onClick={() => router.push('/NovaPlanta/step-2')}
               >
                 Continuar
               </Button>
             </SubmitDiv>
+            <BottomMenu backgroundColor="#FFF"></BottomMenu>
             </Content>
           </ContentFormSM>
     </Main>
