+++ pages/NovaPlanta/step-2.tsx
@@ -4,9 +4,11 @@ import Image from "next/image";
 
 import Main from "../../components/Main";
 
-import { CardSelect, Content, ContentForm, ContentFormSM, DivCards, NovaPlantaHeader, NovaPlantaText, SubmitDiv } from "../../styles/NovaPlanta";
+import { CardSelect, Content, ContentForm, DivCards, NovaPlantaHeader, NovaPlantaText, SubmitDiv } from "../../styles/NovaPlanta";
 import Input from "../../components/Input";
 import Button from "../../components/Button";
+import { useRouter } from "next/router";
+import BottomMenu from "../../components/BottomMenu";
 
 type Inputs = {
   nome: string,
@@ -15,7 +17,7 @@ type Inputs = {
   senha: string,
 };
 
-export default function Step1(props) {
+export default function Step2(props) {
   const {
     register,
     handleSubmit,
@@ -24,59 +26,51 @@ export default function Step1(props) {
   } = useForm<Inputs>();
   
   const onSubmit: SubmitHandler<Inputs> = data => console.log(data);
-  
+  const router = useRouter();
+
   console.log(watch("email"), errors)
   
   return (
     <Main backgroundColor="#FFFFFF" padding="2">
         <NovaPlantaHeader>
+            <Image src="/icone_folha_plantei.svg" alt="Ícone plantei" width={95} height={95} />
           <h1>Sua nova <span>planta</span></h1>
         </NovaPlantaHeader>
-          <ContentFormSM onSubmit={handleSubmit(onSubmit)}>
-            <Content className="contentsm">
+          <ContentForm onSubmit={handleSubmit(onSubmit)}>
+            <Content>
             <div>
-              <NovaPlantaText>Você está em:</NovaPlantaText>
-                <Input
-                  {...register("email", { required: true })}
-                  placeholder="Endereço"
-                  border="1px solid #9AC979"
-                  backgroundColor="#9AC979"
-                  color="#FFF"
-                  width={100}
-                  margin="0"
-                />
-            </div>
-            <div>
-              <NovaPlantaText>Quanto espaço você tem para plantar?</NovaPlantaText>
+              <NovaPlantaText>Selecione onde ela irá ser plantada:</NovaPlantaText>
               <DivCards>
                 <CardSelect>
-                  <Image src="/planta_espaco_pequeno.png" alt="Pequeno" width={95} height={95} />
-                  <p>Pequeno</p>
+                  <Image src="/planta_vaso_plastico.png" alt="Vaso de plástico" width={95} height={95} />
+                  <p>Vaso de plástico</p>
                 </CardSelect>
                 <CardSelect>
-                  <Image src="/planta_espaco_medio.png" alt="Médio" width={95} height={95} />
-                  <p>Médio</p>
+                  <Image src="/planta_vaso_garrafa.png" alt="Garrafa pet" width={95} height={95} />
+                  <p>Garrafa pet</p>
                 </CardSelect>
                 <CardSelect>
-                  <Image src="/planta_espaco_grande.png" alt="Grande" width={95} height={95} />
-                  <p>Grande</p>
+                  <Image src="/planta_vaso_ceramica.png" alt="Vaso de cerâmica" width={95} height={95} />
+                  <p>Vaso de cerâmica</p>
                 </CardSelect>
               </DivCards>
             </div>
             
-            <SubmitDiv className="contentsm">
+            <SubmitDiv>
               <Button
                 backgroundColor="#217B51"
                 width="min-content"
                 color="#FFF"
                 padding="15px 100px"
                 fontSize="18px"
+                onClick={() => router.push('/NovaPlanta/step-3')}
               >
                 Continuar
               </Button>
             </SubmitDiv>
+            <BottomMenu backgroundColor="#FFF" />
             </Content>
-          </ContentFormSM>
+          </ContentForm>
     </Main>
   );
 }
