+++ pages/NovaPlanta/step-5.tsx
@@ -4,9 +4,11 @@ import Image from "next/image";
 
 import Main from "../../components/Main";
 
-import { CardSelect, Content, ContentForm, ContentFormSM, DivCards, NovaPlantaHeader, NovaPlantaText, SubmitDiv } from "../../styles/NovaPlanta";
+import { CardSelect, Content, ContentForm, ContentFormSM, DivCards, NovaPlantaHeader, NovaPlantaText, SubmitDiv, SubmitDivTwoButtons } from "../../styles/NovaPlanta";
 import Input from "../../components/Input";
 import Button from "../../components/Button";
+import { useRouter } from "next/router";
+import BottomMenu from "../../components/BottomMenu";
 
 type Inputs = {
   nome: string,
@@ -24,7 +26,7 @@ export default function Step5(props) {
   } = useForm<Inputs>();
   
   const onSubmit: SubmitHandler<Inputs> = data => console.log(data);
-  
+  const router = useRouter();
   console.log(watch("email"), errors)
   
   return (
@@ -37,13 +39,14 @@ export default function Step5(props) {
                 <Image src="/planta_etapa_final.png" alt="Planta Cadastrada" width={220} height={240} />
                 <h2>1° Passo</h2>
                 <p>Corte a garrafa PET</p>
-                <SubmitDiv>
+                <SubmitDivTwoButtons>
                 <Button
                     backgroundColor="#F3F3F3"
                     width="min-content"
                     color="#217B51"
                     padding="20px 45px"
                     fontSize="18px"
+                    onClick={() => router.push('/NovaPlanta/step-4')}
                 >
                     Voltar
                 </Button>
@@ -53,10 +56,12 @@ export default function Step5(props) {
                     color="#FFF"
                     padding="20px 40px"
                     fontSize="18px"
+                    onClick={() => router.push('/NovaPlanta')}
                 >
                     Finalizar
                 </Button>
-                </SubmitDiv>
+                </SubmitDivTwoButtons>
+                <BottomMenu backgroundColor="#FFF"></BottomMenu>
             </Content>
           </ContentFormSM>
     </Main>
