+++ pages/Profile/index.tsx
@@ -1,9 +1,7 @@
 import { useForm, SubmitHandler } from "react-hook-form";
-import Image from "../../node_modules/next/image";
 
 import Input from "../../components/Input/index";
 import BottomMenu from "../../components/BottomMenu/index";
-import Button from "../../components/Button/index";
 import Main from "../../components/Main";
 
 
@@ -38,7 +36,7 @@ export default function Profile(props) {
         <SpanInfo>Nome</SpanInfo>
         <Input
           {...register("nome", { required: true })}
-          placeholder="Nome completo"
+          placeholder="Jeni Martins"
           border="1px solid #9AC979"
           backgroundColor="#9AC979"
           color="#FFFFFF"
@@ -49,7 +47,7 @@ export default function Profile(props) {
         <SpanInfo>Localização</SpanInfo>
         <Input
           {...register("localizacao", { required: true })}
-          placeholder="Sua cidade"
+          placeholder="São Paulo"
           border="1px solid #9AC979"
           backgroundColor="#9AC979"
           color="#FFFFFF"
@@ -60,7 +58,7 @@ export default function Profile(props) {
         <SpanInfo>Espaço de plantio</SpanInfo>
         <Input
           {...register("espacoPlantio", { required: true })}
-          placeholder="Espaço do plantio selecionado"
+          placeholder="Espaço Médio"
           border="1px solid #9AC979"
           backgroundColor="#9AC979"
           color="#FFFFFF"
